Change the docs and defaults from pixel to sky
Adjust the documentation and default values in the config and in unit
tests to reflect the change in the definition of ``sigmas`` from pixel
coordinates to sky coordinates. The new default values do not have
one-to-one corespondence to previous values but are approximate
conversions assuming a pixel scale of 0.2"/pixel.