Fix the breakage in unit tests
As ``testCorrelatedNoiseError`` tests internal states, this is immune
to change in the definition of units of ``sigmas``. The breakage is a
result of change in the default values. In ``testFlags``, include the
redefinition of ``sigma`` by dividing it by the pixel scale.
Additionally, allow for floating point errors by modifying the
inequality with a small, non-zero tolerance.